review integer types and operations with them
1. All integers are unsigned and use type u64
2. Use plain bit shift operator instead checked bit shift

+++ offchain/compute-server/src/machine/constants.rs
@@ -1,13 +1,13 @@
 use crate::utils::arithmetic;
 
-pub const LEVELS: usize = 3;
-pub const MAX_CYCLE: u8 = 63;
+pub const LEVELS: u64 = 3;
+pub const MAX_CYCLE: u64 = 63;
 
-pub const LOG2_STEP: [u32; 3] = [31, 16, 0];
-pub const HEIGHTS: [u8; 3] = [32, 15, 16];
+pub const LOG2_STEP: [u64; 3] = [31, 16, 0];
+pub const HEIGHTS: [u64; 3] = [32, 15, 16];
 
-pub const LOG2_UARCH_SPAN: u32 = 16;
-pub const UARCH_SPAN: i64 = arithmetic::max_uint(LOG2_UARCH_SPAN);
+pub const LOG2_UARCH_SPAN: u64 = 16;
+pub const UARCH_SPAN: u64 = arithmetic::max_uint(LOG2_UARCH_SPAN);
 
-pub const LOG2_EMULATOR_SPAN: u32 = 47;
-pub const EMULATOR_SPAN: i64 = arithmetic::max_uint(LOG2_EMULATOR_SPAN);
+pub const LOG2_EMULATOR_SPAN: u64 = 47;
+pub const EMULATOR_SPAN: u64 = arithmetic::max_uint(LOG2_EMULATOR_SPAN);
